flatpak: Use default compiler flags.
Using O3 instead of the default O2 isn't really worth it. Using O3
carries risks for compiler bugs which are not that tested, O2 is
the default setting for all major projects. That is not to say that
certain parts might benefit from using -O3, but not a complete
project.