libgeotiff: Use bundled libgeotiff on linux.
Linux is heavily designed to use shared libraries. Trying to use
static libraries within the by definition shared plugin seems to lead to
all sorts of trouble. Instead, just link to to the dynamic, system libs
and bundle them with the plugin.

The cost is the weight of the libraries which makes plugin size
increase. OTOH, this is simple, the code is generic and adding new libs
to bundle is trivial.